+++ packages/components/CHANGELOG.md
@@ -8,9 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+### [2.9.0-beta.1] - 2022-01-27
+
 #### Added
 - `:focus-visible` Content of selected Tab in `Tabs` component gets focus styling
 - improved accessibility of `Text Field Wrapper` and `Textarea Wrapper` when `maxlength` attribute is set
+- `Modal` aria property
+- `Modal` class for slotted elements to make content full-width
+
+#### Changed
+- `Button Pure` and `Link Pure` removed `position: relative` imposition, make sure to **not** override it with `position: static`
+
+#### Fixed
+- `Modal` close button styles when no heading is passed
 
 ### [2.9.0-beta.0] - 2022-01-18